Refactored code. Moved common methods from TableCreator to MetaDataManagerImpl.
