+++ src/main/java/com/custom/orm/metadata/MetaDataManager.java
@@ -19,4 +19,18 @@ public interface MetaDataManager {
     <T> String getTableNameWithoutSchema(Class<T> object);
 
     String getColumnType(Field field);
+
+    String getPrimaryKeyColumnName(Class<?> entityClass);
+
+    String getComposedPrimaryKeyColumnsNames(Class<?> entityClass);
+
+    boolean hasForeignKey(Class<?> entityClass);
+
+    String getForeignKeyColumnName(Class<?> entityClass);
+
+    String getForeignKeyName(Class<?> entityClass);
+
+    String getForeignKeyReferenceClassName(Class<?> entityClass);
+
+    String getForeignKeyReferenceColumnName(Class<?> entityClass);
 }
